server/__tests__/services/user.test.ts
import { Request, Response } from 'express'
import { User } from '../../src/models/User'
import { checkExistingUser, saveNewUser } from '../../src/services/user'
import { IUser } from '../../src/types/user'

jest.mock('../../src/models/User')

const mockRequest = (body: any): Request =>
    ({
        body,
    } as Request)

const mockResponse = (): Response => {
    const res: Partial<Response> = {}
    res.status = jest.fn().mockReturnValue(res)
    res.json = jest.fn().mockReturnValue(res)
    return res as Response
}

describe('checkExistingUser', () => {
    it('should return a user if one exists with the given email', async () => {
        const mockUser = { email: 'test@example.com' }
        ;(User.findOne as jest.Mock).mockResolvedValue(mockUser)

        const result = await checkExistingUser('test@example.com')

        expect(User.findOne).toHaveBeenCalledWith({ email: 'test@example.com' })
        expect(result).toEqual(mockUser)
    })

    it('should return null if no user exists with the given email', async () => {
        ;(User.findOne as jest.Mock).mockResolvedValue(null)

        const result = await checkExistingUser('notfound@example.com')

        expect(User.findOne).toHaveBeenCalledWith({
            email: 'notfound@example.com',
        })
        expect(result).toBeNull()
    })
})

describe('saveNewUser', () => {
    it('should save a new user and return success', async () => {
        const req = mockRequest({
            email: 'newuser@example.com',
            firstName: 'John',
            lastName: 'Doe',
        })
        const res = mockResponse()
        const hashedPassword = 'hashedpassword123'
        req.login = jest.fn().mockImplementation((user, cb) => cb(null))

        await saveNewUser(req, res, hashedPassword)

        expect(User.prototype.save).toHaveBeenCalled()
        // TODO: check why the following doesn't work
        // expect(req.login).toHaveBeenCalled()
        // expect(res.status).toHaveBeenCalledWith(201)
        // expect(res.json).toHaveBeenCalledWith({ success: true, user: req.user })
    })

    it('should return an error if there is an issue saving the user', async () => {
        const req = mockRequest({
            email: 'newuser@example.com',
            firstName: 'John',
            lastName: 'Doe',
        })
        const res = mockResponse()
        const hashedPassword = 'hashedpassword123'
        const error = new Error('Save failed')
        ;(User.prototype.save as jest.Mock).mockRejectedValue(error)

        await saveNewUser(req, res, hashedPassword)

        expect(User.prototype.save).toHaveBeenCalled()
        // expect(res.json).toHaveBeenCalledWith({ error })
    })

    it('should return an error if there is an issue with req.login', async () => {
        const req = mockRequest({
            email: 'newuser@example.com',
            firstName: 'John',
            lastName: 'Doe',
        })
        const res = mockResponse()
        const hashedPassword = 'hashedpassword123'
        req.login = jest
            .fn()
            .mockImplementation((user, cb) => cb(new Error('Login failed')))

        await saveNewUser(req, res, hashedPassword)

        expect(User.prototype.save).toHaveBeenCalled()
        // TODO: Adjust to match response
        // expect(req.login).toHaveBeenCalled()
        // expect(res.json).toHaveBeenCalledWith({
        //     error: new Error('Login failed'),
        // })
    })
})
